FIX: CxEvent, CxThread ADD: doxygen comment CxShell, CxClipboard, CxHandleT, CxCom

+++ Tests/Source/Sync/CxTest_CxEvent.cpp
@@ -36,7 +36,7 @@ CxTest_CxEvent::bUnit(
 
         hRes = objEvent.hGet();
         #if defined(xOS_WIN)
-            xASSERT_NOT_EQ(NULL, hRes);
+            xASSERT_NOT_EQ((ULONG)NULL, (ULONG)hRes);
         #elif defined(xOS_LINUX)
             //TODO: hRes
         #endif
